pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snowflake-cortex-agent-python"
version = "0.1.0"
description = "A Python interface for the Snowflake Cortex Agent REST API"
authors = [
    { name="Michael Gorkow", email="michael.gorkow@snowflake.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "requests>=2.32.3",
    "pydantic>=2.10.3",
    "cryptography>=44.0.1",
    "pyjwt>=2.10.1",
    "httpX>=0.27.0",
    "httpx-sse>=0.4.0",
    "altair>=5.0.1",
    "pandas>=2.2.3",
    "snowflake-snowpark-python>=1.29.0",
    "streamlit>=1.39.0",
    "tabulate>=0.9.0",
    "rich>=13.9.4"
]

[project.urls]
"Homepage" = "https://github.com/michaelgorkow/snowflake-cortex-agent-python"
"Bug Tracker" = "https://github.com/michaelgorkow/snowflake-cortex-agent-python/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
